+++ .editorconfig
@@ -1,11 +1,5 @@
-# Get the plugin for your editor and your
-# tab settings will be set automatically.
-# http://EditorConfig.org
-
-# top-most EditorConfig file
 root = true
 
-# Unix-style newlines with newline ending every file
 [*]
 end_of_line = lf
 insert_final_newline = true
@@ -14,7 +8,3 @@ indent_size = 2
 charset = utf-8
 trim_trailing_whitespace = true
 insert_final_newline = true
-
-# Indentation override for all JS under lib directory
-[*.js]
-indent_size = 4
