+++ .eslintrc
@@ -1,0 +1,6 @@
+{
+  "extends": "airbnb-base",
+  "rules": {
+    "no-console": 0
+  }
+}
