+++ .travis.yml
@@ -7,7 +7,10 @@ os:
   - osx
 
 node_js:
+  - 7
   - 6
-  - 4
-  - '0.12'
-  - '0.11'
+
+cache: yarn
+
+notifications:
+  email: false
