+++ Gruntfile.js
@@ -1,8 +1,8 @@
-'use strict';
+
 
 // This is going away soon.
 // It is still used for the readme.
 // $ grunt repos && grunt readme
 module.exports = function (grunt) {
-    require('load-grunt-tasks')(grunt);
+  require('load-grunt-tasks')(grunt);
 };
