+++ bin/cli.js
@@ -1,9 +1,9 @@
 #!/usr/bin/env node
 
-var isEs2015;
+let isEs2015;
 try {
-    isEs2015 = new Function('() => {}');
+  isEs2015 = new Function('() => {}');
 } catch (e) {
-    isEs2015 = false;
+  isEs2015 = false;
 }
 isEs2015 ? require('../lib/cli') : require('../lib-es5/cli');
