--- lib/index.js
@@ -1,11 +1,0 @@
-'use strict';
-
-const npmCheck = require('./in');
-const createState = require('./state/state');
-
-function init(userOptions) {
-    return createState(userOptions)
-        .then(currentState => npmCheck(currentState));
-}
-
-module.exports = init;
