--- lib/state/debug.js
@@ -1,10 +1,0 @@
-'use strict';
-const chalk = require('chalk');
-
-function debug() {
-    console.log(chalk.green('[npm-check] debug'));
-    console.log.apply(console, arguments);
-    console.log(`${chalk.green('===============================')}`);
-}
-
-module.exports = debug;
