+++ package.json
@@ -7,24 +7,13 @@
     "node": ">=0.11.0"
   },
   "scripts": {
-    "lint": "xo ./lib/*.js",
+    "lint": "eslint --cache --format=node_modules/eslint-formatter-pretty .",
     "test": "npm run lint && ./bin/cli.js || echo Exit Status: $?.",
-    "transpile": "babel lib --out-dir lib-es5",
-    "watch": "babel lib --out-dir lib-es5 --watch",
-    "prepublish": "npm run transpile",
+    "build": "babel src --out-dir lib-es5",
+    "watch": "babel src --out-dir lib-es5 --watch",
+    "prepublish": "npm run build",
     "repos": "grunt repos # required for grunt readme to run",
     "readme": "grunt readme"
-  },
-  "xo": {
-    "space": 4,
-    "rules": {
-      "no-warning-comments": [
-        0
-      ],
-      "global-require": [
-        0
-      ]
-    }
   },
   "bin": {
     "npm-check": "bin/cli.js"
@@ -65,41 +54,44 @@
     "lib-es5"
   ],
   "dependencies": {
-    "babel-runtime": "^6.6.1",
-    "callsite-record": "^3.0.0",
+    "babel-runtime": "^6.22.0",
+    "callsite-record": "^3.2.2",
     "chalk": "^1.1.3",
     "co": "^4.6.0",
-    "depcheck": "^0.6.3",
-    "execa": "^0.2.2",
+    "depcheck": "^0.6.7",
+    "execa": "^0.6.0",
     "giturl": "^1.0.0",
-    "global-modules": "^0.2.0",
-    "globby": "^4.0.0",
-    "inquirer": "^0.12.0",
-    "is-ci": "^1.0.8",
-    "lodash": "^4.7.0",
+    "global-modules": "^0.2.3",
+    "globby": "^6.1.0",
+    "inquirer": "^3.0.1",
+    "is-ci": "^1.0.10",
+    "lodash": "^4.17.4",
     "meow": "^3.7.0",
     "merge-options": "0.0.64",
-    "minimatch": "^3.0.2",
-    "node-emoji": "^1.0.3",
-    "ora": "^0.2.1",
-    "package-json": "^2.0.1",
-    "path-exists": "^2.1.0",
+    "minimatch": "^3.0.3",
+    "node-emoji": "^1.5.1",
+    "ora": "^1.1.0",
+    "package-json": "^2.4.0",
+    "path-exists": "^3.0.0",
     "pkg-dir": "^1.0.0",
-    "semver": "^5.0.1",
-    "semver-diff": "^2.0.0",
+    "semver": "^5.3.0",
+    "semver-diff": "^2.1.0",
     "text-table": "^0.2.0",
-    "throat": "^2.0.2",
-    "update-notifier": "^0.6.3"
+    "throat": "^3.0.0",
+    "update-notifier": "^1.0.3"
   },
   "devDependencies": {
-    "babel-cli": "^6.6.5",
-    "babel-plugin-transform-runtime": "^6.6.0",
-    "babel-preset-es2015": "^6.6.0",
-    "grunt": "^0.4.0",
+    "babel-cli": "^6.22.2",
+    "babel-plugin-transform-runtime": "^6.22.0",
+    "babel-preset-es2017": "^6.22.0",
+    "eslint": "^3.14.0",
+    "eslint-config-airbnb-base": "^11.0.1",
+    "eslint-formatter-pretty": "^1.1.0",
+    "eslint-plugin-import": "^2.2.0",
+    "grunt": "^1.0.1",
     "grunt-cli": "^1.2.0",
-    "grunt-release": "^0.13.0",
-    "grunt-templates-dylang": "^1.0.9",
-    "load-grunt-tasks": "^3.3.0",
-    "xo": "^0.13.0"
+    "grunt-release": "^0.14.0",
+    "grunt-templates-dylang": "^1.0.12",
+    "load-grunt-tasks": "^3.5.2"
   }
 }
